requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-environ>=0.11
django-cors-headers>=4.4
django-ratelimit>=4.1
# NEW for MongoDB Atlas
mongoengine==0.29.1
dnspython>=2.6